Computer_Vision_Challenge/trackmania_yolo/trackmania/src/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pyautogui>=0.9.53
keyboard>=0.13.5
pyyaml>=6.0.1